First attractor prototype
Camera have these options
2D:
 -camera follows curve AND its orientation too
 - camera follows curve but its orientation stays unaffected (this
could be done by projecting curve tangent vector every time camera
moves to a plane which orientation defines how camera is oriented whole
time)
3D:
 -camera follows curve and its orientation
 -camera follows curve and angain to stay stable and feel like natural
camera man style is not rotated with the curve but rotation is
compensated same as before by projection plane..

Camera not being oriened the same way as curve arised new problem -
when camera is not oriented along curve, swiping is also not orinted in
wanted curve swiping direction, so this was solved by projecting
panning gesture difference to tangent of curve on current position
(actualy it is not exacly projecting but I used dot product of swipe
vector and tangent - it is less performance heavy but has the same
result)

First attempt to implement points on curve camera is attracted to.
-> Interval class - interval of attraction area with stop point in the
center
-> splineWalker has its spline but newly also an array of “attraction”
intervals for this spline
-> new state for walker - AttractionZone (besides FreeZone BumpBack,
BumpForward) it helps to drive logic of panning in CurvePanHandler
-> how attractor works very simplified version - CurvePanHandler is
chcecking what is the exact place inertial movement will stop at. if
walker will stop in attractor interval we have to somehow treat it so
it snaps to the center of attractor interval..
-> this attractor prototype works as stand alone attractor = if you lay
down intervals of attractors without any gap (wheel of fortune etc. )
it will perform not very smoothly because in this prototype when walker
gets rejected by first atractor, but lends directly into the attractor
zone after first one, when freemove inertia stops and only after that
second attractor starts moving walker to the center. In this case it
doesnt feel quite right, but I will try to create another prototype
which could cover even this use case (aka wheel of fortune) to feel
awesome like the rest of it.